+++ docs/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @snow-design/docs
 
+## 0.0.22
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.22
+
+## 0.0.21
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.21
+
+## 0.0.20
+
+### Patch Changes
+
+- 修复 Vite 插件 context undefined 报错
+- Updated dependencies
+  - @snow-design/components@0.0.20
+
+## 0.0.19
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.19
+
+## 0.0.18
+
+### Patch Changes
+
+- 修复 Vite 插件
+- Updated dependencies
+  - @snow-design/components@0.0.18
+
+## 0.0.17
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.17
+
+## 0.0.16
+
+### Patch Changes
+
+- 修复 Vite 插件
+- Updated dependencies
+  - @snow-design/components@0.0.16
+
+## 0.0.15
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.15
+
+## 0.0.14
+
+### Patch Changes
+
+- 修复 vite 插件错误
+- Updated dependencies
+  - @snow-design/components@0.0.14
+
+## 0.0.13
+
+### Patch Changes
+
+- 新增 Vite 插件
+- Updated dependencies
+  - @snow-design/components@0.0.13
+
 ## 0.0.12
 
 ### Patch Changes