+++ packages/theme-default/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @snow-design/theme-default
 
+## 0.0.22
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+
+## 0.0.21
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+
+## 0.0.20
+
+### Patch Changes
+
+- 修复 Vite 插件 context undefined 报错
+
+## 0.0.19
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+
+## 0.0.18
+
+### Patch Changes
+
+- 修复 Vite 插件
+
+## 0.0.17
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+
+## 0.0.16
+
+### Patch Changes
+
+- 修复 Vite 插件
+
+## 0.0.15
+
+### Patch Changes
+
+- 修复 Vite 插件错误
+
+## 0.0.14
+
+### Patch Changes
+
+- 修复 vite 插件错误
+
+## 0.0.13
+
+### Patch Changes
+
+- 新增 Vite 插件
+
 ## 0.0.12
 
 ### Patch Changes